+++ es-core/src/main/java/com/izaodao/projects/springboot/elasticsearch/search/match/MatchQuery.java
@@ -47,10 +47,10 @@ public class MatchQuery extends EsQuery implements Serializable {
     private String maxExpansions;
 
     public MatchQuery(String value, String field) {
-        this(value, field, 1.0d);
+        this(value, field, 1.0f);
     }
 
-    public MatchQuery(String value, String field, double boost) {
+    public MatchQuery(String value, String field, float boost) {
         super(boost);
         this.field = field;
         this.value = value;
